GRÁFICAS, Y COMPORBAR SI ES ADMIN

+++ controllers/ClienteController.php
@@ -39,13 +39,15 @@ class ClienteController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new ClienteSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if(isAdmin()){
+            $searchModel = new ClienteSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-            'searchModel'=>$searchModel
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+                'searchModel'=>$searchModel
+            ]);
+        }
     }
 
     /**
@@ -56,9 +58,11 @@ class ClienteController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(isAdmin()){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
     }
 
     /**
@@ -68,19 +72,21 @@ class ClienteController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Cliente();
+        if(isAdmin()){
+            $model = new Cliente();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
-        }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }
     }
 
     /**
@@ -92,15 +98,17 @@ class ClienteController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(isAdmin()){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -112,9 +120,11 @@ class ClienteController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(isAdmin()){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
     }
 
     /**
